Changed return type of `os::pagesize` in stout.
Previously, `os::pagesize` returned a signed integer (`int`), where a
negative value indicated an error. This is problematic: several callers
neglected to check for errors and most callers ultimately want an
unsigned value anyway.

We could make this return `Try<size_t>`, but that seems heavyweight: the
Win32 implementation of this function will never fail and the POSIX
implementation (using `sysconf(3)`) seems unlikely to fail in
practice. Hence, this commit changes `os::pagesize` to return `size_t`.

Review: https://reviews.apache.org/r/51967/

+++ 3rdparty/stout/include/stout/os/freebsd.hpp
@@ -42,7 +42,7 @@ inline Result<Process> process(pid_t pid)
 
   foreach (const kinfo_proc& kinfo, kinfos.get()) {
     if (kinfo.ki_pid == pid) {
-      int pagesize = os::pagesize();
+      size_t pagesize = os::pagesize();
       return Process(kinfo.ki_pid,
                      kinfo.ki_ppid,
                      kinfo.ki_pgid,
@@ -84,7 +84,7 @@ inline Try<Memory> memory()
   }
   memory.total = Bytes(physicalMemory.get());
 
-  const int pageSize = os::pagesize();
+  const size_t pageSize = os::pagesize();
 
   unsigned int freeCount;
   size_t length = sizeof(freeCount);
